# REST APIs Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://todo.swagger.io/v2
  - url: http://todo.swagger.io/v2
...
paths:
...
  /api/films/public/{filmId}/reviews:
    get:
      tags:
        - "films/public/{filmId}/reviews"
      summary:
        "Retrieve the list of all the reviews that have been issued/completed\
        \ for a film, possibly filtered by the invitation Status if the user is the\
        \ film owner"
      description:
        "All the reviews that have been issued/completed for the film with\
        \ ID filmId are retrieved. If the requesting user is the owner of the film,\
        \ he also sees the reviews invitationStatus, otherwise it is not shown. A\
        \ pagination mechanism is implemented to limit the size of messages. This\
        \ operation does not require authentication."
      operationId: getFilmReviews
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered). This query parameter only has effect for\
            \ the film owner."
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "ID of the requested page (if absent, the first page is returned)'"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmIdreviews
    post:
      tags:
        - "films/public/{filmId}/reviews"
      summary: Issue film review to some users
      description:
        The film with ID filmId is assigned to one or more users for review
        and the corresponding reviews are created. The users are specified in the
        review representations in the request body. This operation can only be performed
        by the owner.
      operationId: issueFilmReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description:
          "only the necessary parts of the new film reviews, including\
          \ the users to whom they are issued"
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Review_post"
        required: true
      responses:
        "207":
          description: Multi-Status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/inline_response_207"
                x-content-type: application/json
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (e.g., wrong film id or unknown reviewer id)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviews
  /api/films/public/{filmId}/reviews/invitationStatus:
    get:
      tags:
        - "films/public/{filmId}/reviews/invitationStatus"
      summary:
        "Retrieve the invitationStatus for all reviews that has been issued/completed\
        \ for a film, possibly filtered"
      description:
        The invitationStatus for all the reviews of the film with ID filmID
        are retrieved. This operation can only be executed by the authenticated film
        owner.
      operationId: getReviewsInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered). This query parameter only has effect for\
            \ the film owner."
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "ID of the requested page (if absent, the first page is returned)'"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_2"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsinvitationStatus
  /api/films/public/{filmId}/reviews/{reviewerId}:
    get:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Retrieve a review that has been issued/completed for a film
      description:
        "The review of the film with ID filmID issued to the user with\
        \ ID reviewerId is retrieved. This operation does not require authentication.\
        \ If it is called by the authenticated owner or reviewer, it also returns\
        \ the invitationStatus of the review."
      operationId: getSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_3"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    put:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Complete a review
      description:
        "The review of the film with ID filmId and issued to the user with\
        \ ID reviewerId is completed. This operation only allows setting the \"completed\"\
        \ property to the \"true\" value, and changing the values of the \"reviewDate\"\
        , \"rating\", and \"review\" properties. This operation can be performed only\
        \ by the invited reviewer."
      operationId: updateSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review must be completed
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The properties of the Review object that must be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review_public"
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    delete:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Delete a review invitation
      description:
        "The review of the film with ID filmId and issued to the user with\
        \ ID reviewerId is deleted. This operation can only be performed by the owner,\
        \ and only if the review has not yet been completed by the reviewer."
      operationId: deleteSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review invitation must be deleted
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
    patch:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}"
      summary: Change invitation Status of a review
      description:
        Change the invitation Status of the review of the film with ID
        filmId and issued to the user with ID reviewerId. This operation allows setting
        the "invitationStatus" property of a "PENDING" review to "ACCEPTED" or "REFUSED".
        This operation can be performed only by the invited reviewer.
      operationId: updateSingleReviewInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review must be completed
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The properties of the Review object that must be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review_invitationStatus"
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description:
            "Conflict (attempt to make unallowed change from ACCEPTED/REFUSED\
            \ review, or to PENDING)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerId
  /api/films/public/{filmId}/reviews/{reviewerId}/invitationStatus:
    get:
      tags:
        - "films/public/{filmId}/reviews/{reviewerId}/invitationStatus"
      summary:
        Retrieve the invitationStatus for a review that has been issued/completed
        for a film
      description:
        The invitationStatus for the review of the film with ID filmID
        issued to the user with ID reviewerId is retrieved. This operation can only
        be executed by the authenticated film owner or reviewer.
      operationId: getSingleReviewInvitationStatus
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: reviewerId
          in: path
          description: ID of the user to whom the review has been issued
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review_invitationStatus"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmIdreviewsreviewerIdinvitationStatus
...
components:
  schemas:
...
    Review_full:
      required:
        - completed
        - filmId
        - reviewerId
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description:
            unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
        completed:
          type: boolean
          description: "true if the review has been completed, false otherwise"
          default: false
        reviewDate:
          type: string
          description:
            the date when the review has been completed by the invited
            user
          format: date
        rating:
          maximum: 10
          minimum: 1
          type: integer
          description: the rating the reviewer has given to the film
        review:
          maxLength: 1000
          type: string
          description: a textual description of the review
        invitationStatus:
          type: string
          description: status of the review invitation
          enum:
            - PENDING
            - ACCEPTED
            - REFUSED
        self:
          type: string
          format: uri
      additionalProperties: false
      description:
        schema for full Review data structures including invitationStatus.
        Only visible by film owner and reviewer
    Review_public:
      required:
        - completed
        - filmId
        - reviewerId
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description:
            unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
        completed:
          type: boolean
          description: "true if the review has been completed, false otherwise"
          default: false
        reviewDate:
          type: string
          description:
            the date when the review has been completed by the invited
            user
          format: date
        rating:
          maximum: 10
          minimum: 1
          type: integer
          description: the rating the reviewer has given to the film
        review:
          maxLength: 1000
          type: string
          description: a textual description of the review
        self:
          type: string
          format: uri
      additionalProperties: false
      description:
        "schema for Review data structures, as seen by not logged users\
        \ or logged users which are not film owner or reviewer, so without the invitationStatus"
    Review_post:
      required:
        - filmId
        - reviewerId
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description:
            unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
      additionalProperties: false
      description:
        Optimized JSON Schema to represent just the relevant Review fields
        when it needs to be created by the film owner
    Review_invitationStatus:
      required:
        - filmId
        - invitationStatus
        - reviewerId
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        filmId:
          type: integer
          description:
            unique identifier of the film for which a review invitation
            has been issued
        reviewerId:
          type: integer
          description: unique identifier of the user who has received the review invitation
        invitationStatus:
          type: string
          description: status of the review invitation
          enum:
            - PENDING
            - ACCEPTED
            - REFUSED
        self:
          type: string
          format: uri
      additionalProperties: false
      description: JSON Schema for the invitation status of a Review
      example:
        reviewerId: 5
        $schema: $schema
        filmId: 5
        self: http://example.com/aeiou
        invitationStatus: "PENDING"
...
    inline_response_200_1:
      type: object
      properties:
        totalPages:
          type: integer
        currentPage:
          type: integer
        totalItems:
          type: integer
        reviews:
          type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/Review_full"
              - $ref: "#/components/schemas/Review_public"
        next:
          type: string
          format: uri
      example:
        next: http://example.com/aeiou
        totalItems: 1
        reviews:
          - ""
          - ""
        totalPages: 0
        currentPage: 6
    inline_response_207:
      type: object
      properties:
        responseStatus:
          type: integer
        content:
          oneOf:
            - $ref: "#/components/schemas/Review_full"
            - type: object
              properties:
                error:
                  type: string
      example:
        responseStatus: 0
        content: ""
    inline_response_200_2:
      type: object
      properties:
        totalPages:
          type: integer
        currentPage:
          type: integer
        totalItems:
          type: integer
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review_invitationStatus"
        next:
          type: string
          format: uri
      example:
        next: http://example.com/aeiou
        totalItems: 1
        reviews:
          - reviewerId: 5
            $schema: $schema
            filmId: 5
            self: http://example.com/aeiou
            invitationStatus: "PENDING"
          - reviewerId: 5
            $schema: $schema
            filmId: 5
            self: http://example.com/aeiou
            invitationStatus: "PENDING"
        totalPages: 0
        currentPage: 6
    inline_response_200_3:
      oneOf:
        - $ref: "#/components/schemas/Review_full"
        - $ref: "#/components/schemas/Review_public"
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie